+++ test/_helper.js
@@ -1,9 +1,14 @@
 import _ from 'underscore';
 import sinon from 'sinon';
 import * as utils from '../build/utils';
+import * as requestQueues from '../build/request_queues';
 import ApifyClientError, { REQUEST_FAILED_ERROR_TYPE, REQUEST_FAILED_ERROR_MESSAGE } from '../build/apify_error';
 
 let requestMock;
+
+export const DEFAULT_RATE_LIMIT_ERRORS = new Array(
+    Math.max(requestQueues.REQUEST_ENDPOINTS_EXP_BACKOFF_MAX_REPEATS, utils.EXP_BACKOFF_MAX_REPEATS),
+).fill(0);
 
 export const mockRequest = () => {
     requestMock = sinon.mock(utils, 'requestPromise');
@@ -13,7 +18,7 @@ export const newEmptyStats = () => {
     return {
         calls: 0,
         requests: 0,
-        rateLimitErrors: [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+        rateLimitErrors: [...DEFAULT_RATE_LIMIT_ERRORS],
     };
 };