+++ test/utils.js
@@ -12,7 +12,7 @@ import ApifyClientError, {
     INVALID_PARAMETER_ERROR_TYPE_V2,
     NOT_FOUND_STATUS_CODE,
 } from '../build/apify_error';
-import { newEmptyStats } from './_helper';
+import { newEmptyStats, DEFAULT_RATE_LIMIT_ERRORS } from './_helper';
 import * as utils from '../build/utils';
 
 const { CLIENT_USER_AGENT, CONTENT_TYPE_JSON_HEADER } = utils;
@@ -113,7 +113,7 @@ describe('utils.requestPromise()', () => {
                     calls: 1,
                     requests: 1,
                 });
-                expect(stats.rateLimitErrors).to.be.eql([0, 0, 0, 0, 0, 0, 0, 0, 0, 0]);
+                expect(stats.rateLimitErrors).to.be.eql(DEFAULT_RATE_LIMIT_ERRORS);
                 stub.restore();
             });
     });
@@ -178,7 +178,7 @@ describe('utils.requestPromise()', () => {
                     calls: 1,
                     requests: 8,
                 });
-                expect(stats.rateLimitErrors).to.be.eql([0, 0, 0, 0, 0, 0, 0, 0, 0, 0]);
+                expect(stats.rateLimitErrors).to.be.eql(DEFAULT_RATE_LIMIT_ERRORS);
                 stub.restore();
             });
     });
@@ -232,6 +232,7 @@ describe('utils.requestPromise()', () => {
         const expectedBody = 'foo-bar';
 
         let iteration = 0;
+        const rateLimitedIterations = 5;
 
         const stub = sinon
             .stub(request, method.toLowerCase())
@@ -243,22 +244,25 @@ describe('utils.requestPromise()', () => {
                     headers: { 'User-Agent': CLIENT_USER_AGENT },
                 }));
                 iteration++;
-                if (iteration < 6) return Promise.resolve({ statusCode: 429 });
+                if (iteration <= rateLimitedIterations) return Promise.resolve({ statusCode: 429 });
                 return Promise.resolve({ body: expectedBody });
             });
 
         const stats = newEmptyStats();
 
+        const expectedRateLimitErrors = [...DEFAULT_RATE_LIMIT_ERRORS];
+        for (let i = 0; i < rateLimitedIterations; i++) expectedRateLimitErrors[i] = 1;
+
         return utils
             .requestPromise(opts, stats)
             .then((body) => {
                 expect(body).to.be.eql(expectedBody);
-                expect(iteration).to.be.eql(6);
+                expect(iteration).to.be.eql(rateLimitedIterations + 1);
                 expect(stats).to.include({
                     calls: 1,
-                    requests: 6,
+                    requests: rateLimitedIterations + 1,
                 });
-                expect(stats.rateLimitErrors).to.be.eql([1, 1, 1, 1, 1, 0, 0, 0, 0, 0]);
+                expect(stats.rateLimitErrors).to.be.eql(expectedRateLimitErrors);
                 stub.restore();
             });
     });
@@ -484,7 +488,7 @@ describe('utils.requestPromise()', () => {
                     calls: 1,
                     requests: 1,
                 });
-                expect(stats.rateLimitErrors).to.be.eql([0, 0, 0, 0, 0, 0, 0, 0, 0, 0]);
+                expect(stats.rateLimitErrors).to.be.eql(DEFAULT_RATE_LIMIT_ERRORS);
                 stub.restore();
             });
     });
@@ -515,7 +519,7 @@ describe('utils.requestPromise()', () => {
                     calls: 1,
                     requests: 1,
                 });
-                expect(stats.rateLimitErrors).to.be.eql([0, 0, 0, 0, 0, 0, 0, 0, 0, 0]);
+                expect(stats.rateLimitErrors).to.be.eql(DEFAULT_RATE_LIMIT_ERRORS);
                 stub.restore();
             });
     });